Moved around some files, and started the actual stream
