+++ generate/generate.js
