+++ stream/stream.js
@@ -1,0 +1,5 @@
+
+
+module.exports = (path, config) => {
+  console.log('Starting Stream!');
+}
